My solution to this quiz
I decided to solve the quiz in order to challenge my self. The problem is generalized for strings of arbitrary length, so is not optimized for the length of the strings contained inside the input file. 

Please provide me a feedback when you have some spare time. 

All the best
A.G.